Add home page to routes list. - Add icon to side and nav bars
