+++ client/src/constants/index.js
@@ -5,31 +5,37 @@ import {
   payment,
   profile,
   withdraw,
+  home,
 } from '../assets';
 
 export const navlinks = [
   {
+    name: 'home',
+    imgUrl: home,
+    link: '/',
+  },
+  {
     name: 'dashboard',
     imgUrl: dashboard,
-    link: '/',
+    link: '/campaigns',
   },
   {
     name: 'campaign',
     imgUrl: createCampaign,
     link: '/create-campaign',
   },
-  {
-    name: 'payment',
-    imgUrl: payment,
-    link: '/',
-    disabled: true,
-  },
-  {
-    name: 'withdraw',
-    imgUrl: withdraw,
-    link: '/',
-    disabled: true,
-  },
+  // {
+  //   name: 'payment',
+  //   imgUrl: payment,
+  //   link: '/',
+  //   disabled: true,
+  // },
+  // {
+  //   name: 'withdraw',
+  //   imgUrl: withdraw,
+  //   link: '/',
+  //   disabled: true,
+  // },
   {
     name: 'profile',
     imgUrl: profile,
