+++ client/src/pages/Campaigns.jsx
@@ -1,0 +1,33 @@
+import React, { useState, useEffect } from 'react';
+import { DisplayCampaigns } from '../components';
+import { useStateContext } from '../context';
+
+const Campaigns = () => {
+  const [isLoading, setIsLoading] = useState(false);
+  const [campaigns, setCampaigns] = useState([]);
+
+  const { address, contract, getCampaigns } = useStateContext();
+
+  const fetchCampaigns = async () => {
+    setIsLoading(true);
+    const data = await getCampaigns();
+    setCampaigns(data);
+    setIsLoading(false);
+  };
+
+  useEffect(() => {
+    if (contract) fetchCampaigns();
+  }, [address, contract]);
+
+  return (
+    <div>
+      <DisplayCampaigns
+        title="All Campaigns"
+        isLoading={isLoading}
+        campaigns={campaigns}
+      />
+    </div>
+  );
+};
+
+export default Campaigns;
