+++ client/src/pages/Home.jsx
@@ -1,33 +1,41 @@
-import React, { useState, useEffect } from 'react';
-import { DisplayCampaigns } from '../components';
-import { useStateContext } from '../context';
+// import React, { useState, useEffect } from 'react';
+// import { DisplayCampaigns } from '../components';
+// import { useStateContext } from '../context';
+
+// const Home = () => {
+//   const [isLoading, setIsLoading] = useState(false);
+//   const [campaigns, setCampaigns] = useState([]);
+
+//   const { address, contract, getCampaigns } = useStateContext();
+
+//   const fetchCampaigns = async () => {
+//     setIsLoading(true);
+//     const data = await getCampaigns();
+//     setCampaigns(data);
+//     setIsLoading(false);
+//   };
+
+//   useEffect(() => {
+//     if (contract) fetchCampaigns();
+//   }, [address, contract]);
+
+//   return (
+//     <div>
+//       <DisplayCampaigns
+//         title="All Campaigns"
+//         isLoading={isLoading}
+//         campaigns={campaigns}
+//       />
+//     </div>
+//   );
+// };
+
+// export default Home;
+
+import React from 'react';
 
 const Home = () => {
-  const [isLoading, setIsLoading] = useState(false);
-  const [campaigns, setCampaigns] = useState([]);
-
-  const { address, contract, getCampaigns } = useStateContext();
-
-  const fetchCampaigns = async () => {
-    setIsLoading(true);
-    const data = await getCampaigns();
-    setCampaigns(data);
-    setIsLoading(false);
-  };
-
-  useEffect(() => {
-    if (contract) fetchCampaigns();
-  }, [address, contract]);
-
-  return (
-    <div>
-      <DisplayCampaigns
-        title="All Campaigns"
-        isLoading={isLoading}
-        campaigns={campaigns}
-      />
-    </div>
-  );
+  return <div>Home</div>;
 };
 
 export default Home;
