+++ client/src/pages/index.js
@@ -1,4 +1,5 @@
 export { default as Home } from './Home';
+export { default as Campaigns } from './Campaigns';
 export { default as Profile } from './Profile';
 export { default as CampaignDetails } from './CampaignDetails';
 export { default as CreateCampaign } from './CreateCampaign';
